+++ lib/ui/health/Covid19TestLocations.dart
@@ -16,6 +16,7 @@
 
 import 'dart:collection';
 
+import 'package:flutter_html/style.dart';
 import 'package:illinois/service/Analytics.dart';
 import 'package:illinois/service/NativeCommunicator.dart';
 import 'package:illinois/utils/Utils.dart';
@@ -28,6 +29,7 @@ import 'package:illinois/service/Storage.dart';
 import 'package:illinois/ui/widgets/HeaderBar.dart';
 import 'package:illinois/service/Styles.dart';
 import 'package:illinois/service/LocationServices.dart';
+import 'package:sprintf/sprintf.dart';
 
 class Covid19TestLocationsPanel extends StatefulWidget {
   _Covid19TestLocationsPanelState createState() => _Covid19TestLocationsPanelState();
@@ -80,7 +82,7 @@ class _Covid19TestLocationsPanelState extends State<Covid19TestLocationsPanel>{
   @override
   Widget build(BuildContext context) {
     int itemsLength = _locations?.length ?? 0;
-    itemsLength+= 2; // for dropdowns
+    itemsLength+= 3; // for dropdowns + 1 for empty
     return Scaffold(
       backgroundColor: Styles().colors.background,
       appBar: SimpleHeaderBarWithBack(
@@ -93,22 +95,38 @@ class _Covid19TestLocationsPanelState extends State<Covid19TestLocationsPanel>{
         ? Center(child: CircularProgressIndicator(),)
         : Padding(
           padding: const EdgeInsets.symmetric(horizontal: 16),
-          child: itemsLength>0? ListView.builder(
-            itemCount: itemsLength,
-            itemBuilder: (BuildContext context, int index) {
-              if(index == 0)
-                return _buildCountyField();
-              if(index == 1)
-                return _buildProviderField();
+          child:
+              itemsLength>0? ListView.builder(
+                itemCount: itemsLength,
+                itemBuilder: (BuildContext context, int index) {
+                  if(index == 0)
+                    return _buildCountyField();
+                  if(index == 1)
+                    return _buildProviderField();
+                  if(index == 2)
+                    return _buildEmpty();
 
-              index -= 2; //for dropdowns
-              HealthServiceLocation location = (_locations?.isNotEmpty ?? false)? _locations[index] : null;
-              //double distance = _locationData!=null && location!=null? AppLocation.distance(location.latitude, location.longitude, _locationData.latitude, _locationData.longitude) : 0;
-              return _TestLocation(testLocation: location, /*distance: distance,*/);
-            },
-          ) : Container(),
+                  index -= 3; //for dropdowns
+                  HealthServiceLocation location = (_locations?.isNotEmpty ?? false)? _locations[index] : null;
+                  //double distance = _locationData!=null && location!=null? AppLocation.distance(location.latitude, location.longitude, _locationData.latitude, _locationData.longitude) : 0;
+                  return location!=null? _TestLocation(testLocation: location, /*distance: distance,*/): Container();
+                },
+              ) :Container(),
         ),
     );
+  }
+
+  Widget _buildEmpty(){
+    return _locations.isEmpty?
+    Row(children: [
+      Expanded(child: Container(
+          child: Center(child:
+          Text(sprintf(Localization().getStringEx("panel.covid19_test_locations.empty.title", "No test location available for %s provider"),['${_selectedProviderItem?.provider?.name ?? "All"}']),
+            style: TextStyle(color: Styles().colors.textSurface, fontSize: 16, fontFamily: Styles().fontFamilies.regular), textAlign: TextAlign.center,
+          ),)
+      ))
+    ],)
+     : Container();
   }
 
   void _sortLocations() async{
@@ -339,6 +357,16 @@ class _TestLocation extends StatelessWidget{
         (waitTimeColor == HealthLocationWaitTimeColor.green);
     String waitTimeText = Localization().getStringEx('panel.covid19_test_locations.wait_time.label', 'Wait Time') +
         (isWaitTimeAvailable ? '' : (' ' + Localization().getStringEx('panel.covid19_test_locations.wait_time.unavailable', 'Unavailable')));
+    String waitTimeStatusText ="";
+    if(isWaitTimeAvailable){
+      if(waitTimeColor == HealthLocationWaitTimeColor.red){
+        waitTimeStatusText = Localization().getStringEx('panel.covid19_test_locations.wait_time.status.label.red', 'Long');
+      } else if(waitTimeColor == HealthLocationWaitTimeColor.yellow){
+        waitTimeStatusText = Localization().getStringEx('panel.covid19_test_locations.wait_time.status.label.yellow', 'Medium');
+      } else if(waitTimeColor == HealthLocationWaitTimeColor.green){
+        waitTimeStatusText = Localization().getStringEx('panel.covid19_test_locations.wait_time.status.label.green', 'Short');
+      }
+    }
     return
       Semantics(button: false, container: true, child:
         Container(
@@ -419,6 +447,17 @@ class _TestLocation extends StatelessWidget{
                             height: 16,
                             decoration: BoxDecoration(color: HealthServiceLocation.waitTimeColorHex(waitTimeColor), shape: BoxShape.circle),
                           ),
+                        ),
+                        Padding(
+                          padding: EdgeInsets.only(right: 8),
+                          child: Text(
+                            waitTimeStatusText,
+                            style: TextStyle(
+                              fontFamily: Styles().fontFamilies.regular,
+                              fontSize: 16,
+                              color: Styles().colors.textSurface,
+                            ),
+                          )
                         ),
                         Text(
                           waitTimeText,